backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
SQLAlchemy==2.0.28
python-dotenv==1.0.1 